tradecore/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
stripe==11.3.0
httpx==0.27.2
redis[hiredis]==5.2.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
slowapi==0.1.9
authlib==1.3.2
itsdangerous==2.2.0
email-validator==2.2.0
sendgrid==6.11.0
structlog==24.4.0
websockets==13.1
python-telegram-bot==21.7
apscheduler==3.10.4
